fixes and changelog
