+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- (fix) XAP-188 packages/i18n 1.0.3 packages/ui-helpers 1.0.23 pages/portfolio 1.1.15 | add token to wallet button
+
 - (fix) XAP-110 pages/portfolio 1.1.13 | single token representation: remove test
 
 - (fix) XAP-109 pages/portfolio 1.1.12 | single token representation: potfolio responsive
