+++ packages/i18n/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@evmosapps/i18n",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "private": true,
   "type": "module",
   "devDependencies": {
