+++ pages/portfolio/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@evmosapps/portfolio-page",
-  "version": "1.1.14",
+  "version": "1.1.15",
   "private": true,
   "type": "module",
   "main": "./src/index.tsx",
